requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.3.0
pyyaml>=6.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
langchain-mistralai>=0.0.1
langchain-groq>=0.0.1 